Consolidate search-related logging
